+++ source/gui/ITEXT.cpp
@@ -37,7 +37,7 @@ namespace gui
 			WCHAR* s = (WCHAR*)GlobalLock(h);
 			render::IRenderTextNew * t = (render::IRenderTextNew*)m_pRenderFrame;
 			m_wsText.insert(t->getCaretPos(), s);
-			t->setCaretPos(t->getCaretPos() + wcslen(s), true);
+			t->setCaretPos(t->getCaretPos() + (int)wcslen(s), true);
 			GlobalUnlock(h);
 			CloseClipboard();
 		}
